% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChinaFlux-package.R
\docType{package}
\name{ChinaFlux-package}
\alias{ChinaFlux}
\alias{ChinaFlux-package}
\title{ChinaFlux: Tools for ChinaFlux}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
