% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_Soil.R
\name{interp_soil_profile}
\alias{interp_soil_profile}
\title{垂直插值土壤剖面数据}
\usage{
interp_soil_profile(df, var_prefix = "TS")
}
\arguments{
\item{df}{包含观测数据的数据框}

\item{var_prefix}{变量前缀，如 "TS" 或 "SM"}
}
\value{
返回 ntime x 5 的矩阵，列名为 Prefix_L1 到 Prefix_L5
}
\description{
垂直插值土壤剖面数据
}
