% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_Soil.R
\name{process_site_soil}
\alias{process_site_soil}
\title{批量处理站点的TS和SM，并合并回原数据框}
\usage{
process_site_soil(df)
}
\arguments{
\item{df}{站点原始数据框}
}
\value{
拼接了 TS_L1~L5 和 SM_L1~L5 的新数据框
}
\description{
批量处理站点的TS和SM，并合并回原数据框
}
